Move all the binaries and respective files into cmd/
Keeps all binaries and their support files grouped nicely instead of
littered randomly.
