+++ cmd/tink-cli/cmd/hardware/all.go
