+++ cmd/tink-cli/cmd/hardware/id.go
