+++ cmd/tink-cli/cmd/hardware/watch.go
