+++ cmd/tink-cli/cmd/template/commands.go
