+++ cmd/tink-cli/cmd/template/delete.go
