+++ cmd/tink-cli/cmd/template/get.go
