+++ cmd/tink-cli/cmd/template/list.go
