+++ cmd/tink-cli/cmd/workflow/delete.go
