+++ cmd/tink-cli/sample.tmpl
