+++ cmd/tink-server/deploy/db/docker-entrypoint-initdb.d/tinkerbell-init.sql
